// application/code/book-my-show/src/application/models/product/EwSystem.Interface.ts
/**
 * Author: Brian Brown
 * Date: June 22nd, 2020
 * Description: Used to hold Platform EW System data
 *
 * Change Log:
 *
 * Date         Developer           Description
 * ---------    -----------------   -------------------------------------------------
 *
 */
interface EwSystem {
    name: string;
    docType: string;
    doc: string;
    loa: string;
}

export { EwSystem };